Read product data function
created a function that reads data from the file based off of an inputted product id. this id is translated into an index (product id - 1) and is then used to calculate the number of bytes to offset seek with. then it store the data read into a struct and is returned